requirements.txt
langchain==0.3.7
langchain-community==0.3.7
langchain-core==0.3.19
langchain-text-splitters==0.3.2
langchain-groq==0.2.1
langchain-huggingface==0.1.2
chromadb
sentence-transformers
streamlit
datasets
python-dotenv
fastapi
uvicorn
requests
tqdm
rank_bm25
